test(type): keep the version tests in one file
The split into a separate wire test file was for a codec move that is not
happening yet, so the domain and wire tests sit together like every other type
in the package.
